design: refine UI/UX of user profile forms
