+++ app/(dashboard)/_components/forms/notification-form.tsx
@@ -2,13 +2,11 @@
 
 import * as z from "zod";
 import toast from "react-hot-toast";
-import Link from "next/link";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 
 import { Button } from "@/components/ui/button";
 import { Switch } from "@/components/ui/switch";
-import { Checkbox } from "@/components/ui/checkbox";
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import {
   Form,
@@ -21,14 +19,16 @@ import {
 } from "@/components/ui/form";
 
 const notificationsFormSchema = z.object({
-  notif_level: z.enum(["all", "level-change", "none"], {
+  notif_level: z.enum(["all", "profile", "none"], {
     required_error: "⚠ You need to select a notification type.",
   }),
+  
   communication_emails: z.boolean().default(false).optional(),
-  push_notifications: z.boolean().default(false).optional(),
-  marketing_emails: z.boolean().default(false).optional(),
-  log_events: z.boolean().default(false).optional(),
-  mobile_diff: z.boolean().default(false).optional(),
+  marketing_notifs: z.boolean().default(false).optional(),
+  newsletter_notifs: z.boolean().default(false).optional(),
+  
+  push_notifs: z.boolean().default(false).optional(),
+  mobile_notifs: z.boolean().default(false).optional(),
 });
 type NotificationsFormValues = z.infer<typeof notificationsFormSchema>;
 
@@ -38,9 +38,8 @@ const defaultValues: Partial<NotificationsFormValues> = {
   // 👇 dev testing
   notif_level: "all",
   communication_emails: false,
-  marketing_emails: false,
-  push_notifications: false,
-  log_events: false,
+  marketing_notifs: false,
+  push_notifs: false,
 };
 
 export function NotificationsForm() {
@@ -73,8 +72,11 @@ export function NotificationsForm() {
           control={form.control}
           name="notif_level"
           render={({ field }) => (
-            <FormItem className="space-y-3">
-              <FormLabel>Notify me about...</FormLabel>
+            <FormItem className="space-y-3 py-3">
+              <FormLabel className="mb-4 text-xl font-medium">
+                Live notifications
+              </FormLabel>
+              <FormDescription></FormDescription>
               <FormControl>
                 <RadioGroup
                   onValueChange={field.onChange}
@@ -85,9 +87,7 @@ export function NotificationsForm() {
                     <FormControl>
                       <RadioGroupItem value="all" />
                     </FormControl>
-                    <FormLabel className="font-normal">
-                      All profile updates
-                    </FormLabel>
+                    <FormLabel className="font-normal">All updates</FormLabel>
                   </FormItem>
 
                   <FormItem className="flex items-center space-x-3 space-y-0">
@@ -95,7 +95,7 @@ export function NotificationsForm() {
                       <RadioGroupItem value="level-change" />
                     </FormControl>
                     <FormLabel className="font-normal">
-                      Only level changes
+                      Profile updates
                     </FormLabel>
                   </FormItem>
 
@@ -103,7 +103,7 @@ export function NotificationsForm() {
                     <FormControl>
                       <RadioGroupItem value="none" />
                     </FormControl>
-                    <FormLabel className="font-normal">Nothing</FormLabel>
+                    <FormLabel className="font-normal">No updates</FormLabel>
                   </FormItem>
                 </RadioGroup>
               </FormControl>
@@ -138,7 +138,7 @@ export function NotificationsForm() {
             />
             <FormField
               control={form.control}
-              name="marketing_emails"
+              name="marketing_notifs"
               render={({ field }) => (
                 <FormItem className="flex flex-row items-center justify-between rounded-lg border p-4">
                   <div className="space-y-0.5">
@@ -160,43 +160,21 @@ export function NotificationsForm() {
             />
             <FormField
               control={form.control}
-              name="push_notifications"
+              name="newsletter_notifs"
               render={({ field }) => (
                 <FormItem className="flex flex-row items-center justify-between rounded-lg border p-4">
                   <div className="space-y-0.5">
                     <FormLabel className="text-base">
-                      Push notifications
+                      Newsletter emails
                     </FormLabel>
                     <FormDescription>
-                      Receive push notifications on your device.
+                      Subscirbe to our quarterly open source community letter.
                     </FormDescription>
                   </div>
                   <FormControl>
                     <Switch
                       checked={field.value}
                       onCheckedChange={field.onChange}
-                    />
-                  </FormControl>
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="log_events"
-              render={({ field }) => (
-                <FormItem className="flex flex-row items-center justify-between rounded-lg border p-4">
-                  <div className="space-y-0.5">
-                    <FormLabel className="text-base">Event Logs</FormLabel>
-                    <FormDescription>
-                      Receive live toast notifcations of events.
-                    </FormDescription>
-                  </div>
-                  <FormControl>
-                    <Switch
-                      checked={field.value}
-                      onCheckedChange={field.onChange}
-                      // disabled
-                      // aria-readonly
                     />
                   </FormControl>
                 </FormItem>
@@ -204,29 +182,54 @@ export function NotificationsForm() {
             />
           </div>
         </div>
-        <FormField
-          control={form.control}
-          name="mobile_diff"
-          render={({ field }) => (
-            <FormItem className="flex flex-row items-start space-x-3 space-y-0">
-              <FormControl>
-                <Checkbox
-                  checked={field.value}
-                  onCheckedChange={field.onChange}
-                />
-              </FormControl>
-              <div className="space-y-1 leading-none">
-                <FormLabel>
-                  Use different settings for my mobile devices
-                </FormLabel>
-                <FormDescription>
-                  You can manage your mobile notifications in the{" "}
-                  <Link href="/examples/forms">mobile settings</Link> page.
-                </FormDescription>
-              </div>
-            </FormItem>
-          )}
-        />
+        <h3 className="mb-4 text-lg font-medium">Mobile Notifications</h3>
+        <div className="space-y-4">
+          <FormField
+            control={form.control}
+            name="push_notifs"
+            render={({ field }) => (
+              <FormItem className="flex flex-row items-center justify-between rounded-lg border p-4">
+                <div className="space-y-0.5">
+                  <FormLabel className="text-base">
+                    Push notifications
+                  </FormLabel>
+                  <FormDescription>
+                    Receive push notifications on your device.
+                  </FormDescription>
+                </div>
+                <FormControl>
+                  <Switch
+                    checked={field.value}
+                    onCheckedChange={field.onChange}
+                  />
+                </FormControl>
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mobile_notifs"
+            render={({ field }) => (
+              <FormItem className="flex flex-row items-center justify-between rounded-lg border p-4">
+                <div className="space-y-0.5">
+                  <FormLabel className="text-base">
+                    Different settings for mobile
+                  </FormLabel>
+                  <FormDescription>
+                    Manage how you receive notifications on your mobile devices.
+                  </FormDescription>
+                </div>
+                <FormControl>
+                  <Switch
+                    checked={field.value}
+                    onCheckedChange={field.onChange}
+                  />
+                </FormControl>
+              </FormItem>
+            )}
+          />
+        </div>
+
         <Button
           type="submit"
           variant={"devfill"}
