+++ app/(dashboard)/_components/forms/profile-form.tsx
@@ -1,19 +1,19 @@
 "use client";
 
 import * as z from "zod";
+import toast from "react-hot-toast";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { UseFormSetValue, useForm } from "react-hook-form";
-
+import { useForm } from "react-hook-form";
 import { cn } from "@/lib/utils";
-import { Slider } from "@/components/ui/slider";
+
+import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
+import { Switch } from "@/components/ui/switch";
 import { Textarea } from "@/components/ui/textarea";
-import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
-import { CheckCheckIcon, CheckIcon } from "lucide-react";
 import {
   Popover,
-  PopoverContent,
   PopoverTrigger,
+  PopoverContent,
 } from "@/components/ui/popover";
 import {
   Command,
@@ -31,46 +31,59 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-
 import {
-  careerList,
-  programmingLanguagesList,
-  skillsList,
-} from "@/constants/navigation-index";
+  CheckCheckIcon,
+  CheckIcon,
+  Github,
+  LinkIcon,
+  Linkedin,
+  Medal,
+  Trophy,
+} from "lucide-react";
 
-import toast from "react-hot-toast";
+
+import { countriesList, languages } from "@/constants/userforms-index";
 
 const profileFormSchema = z.object({
-  career_title: z.string({
-    required_error: "⚠ Please pick your career .",
-  }),
-  programming_lang: z.string({
-    required_error: "⚠ Please pick a language .",
-  }),
   bio: z.string().max(160).min(4).optional(),
-  career_level: z.number().min(0, "⚠ Please set your level"),
-  experience_level: z.number().min(0, "⚠ Please set your level"),
-  skills_list: z.array(z.string()),
+
+  country: z.string({
+    required_error: "⚠ Please pick your country.",
+  }),
+  language: z.string({
+    required_error: "⚠ Please select a language.",
+  }),
+  urls: z
+    .object({
+      github: z.string().url().optional(),
+      linkedin: z.string().url().optional(),
+      website: z.string().url().optional(),
+    })
+    .refine((data) => Object.values(data).some(Boolean), {
+      message: "⚠ At least one social media profile is required.",
+    }),
+  projects: z
+    .object({
+      capstone: z.string().url().optional(),
+      additional: z.string().url().optional(),
+    }),
+  ztm_student: z.boolean().default(false).optional(),
+  star_mentor: z.boolean().default(false).optional(),
 });
+
 type ProfileFormValues = z.infer<typeof profileFormSchema>;
 
 const defaultValues: Partial<ProfileFormValues> = {
   // 🎯 to-do-list
   // - will be a database / API call
-  // career_title: {user.name},
-  // bio: {user.bio},
 };
 
 export function ProfileForm() {
   const form = useForm<ProfileFormValues>({
     resolver: zodResolver(profileFormSchema),
     defaultValues,
+    mode: "onChange",
   });
-
-  const handleSkillList = (selectedSkill: string, prevSkills: string[]) => {
-    // 🎯 to-do-list
-    //-I still need to figure out a way to manage the data as form data
-  };
 
   function onSubmit(data: ProfileFormValues) {
     // 🎯 to-do-list
@@ -84,279 +97,20 @@ export function ProfileForm() {
 
   return (
     <Form {...form}>
-      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8 w-full">
-        {/* CAREER AND LANGUAGE */}
-        <div className="flex flex-col justify-center items-center md:justify-normal md:flex-row gap-5 md:gap-24">
-          <FormField
-            control={form.control}
-            name="career_title"
-            render={({ field }) => (
-              <FormItem className="flex flex-col rounded-lg border p-4 space-y-1 w-[270px] ">
-                <FormLabel>Career Title</FormLabel>
-                <FormDescription>
-                  Pick your current or dream title.
-                </FormDescription>
-                <Popover>
-                  <PopoverTrigger asChild>
-                    <FormControl>
-                      <Button
-                        variant="outline"
-                        role="combobox"
-                        className={cn(
-                          "w-[200px] justify-between",
-                          !field.value &&
-                            "text-muted-foreground text-devready-green"
-                        )}
-                      >
-                        {field.value
-                          ? careerList.find(
-                              (language) => language.label === field.value
-                            )?.label
-                          : "Select language"}
-                        <CheckCheckIcon className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-                      </Button>
-                    </FormControl>
-                  </PopoverTrigger>
-                  <PopoverContent className="w-[200px] p-0">
-                    <Command className=" overflow-y-auto max-h-[420px]">
-                      <CommandInput placeholder="Search language..." />
-                      <CommandEmpty>No language found.</CommandEmpty>
-                      <CommandGroup className="overflow-y-auto max-h-[300px]">
-                        {careerList.map((career) => (
-                          <CommandItem
-                            value={career.label}
-                            key={career.label}
-                            onSelect={() => {
-                              form.setValue("career_title", career.label);
-                            }}
-                          >
-                            <CheckIcon
-                              className={cn(
-                                "mr-2 h-4 w-4",
-                                career.label === field.value
-                                  ? "opacity-100"
-                                  : "opacity-0"
-                              )}
-                            />
-                            {career.label}
-                          </CommandItem>
-                        ))}
-                      </CommandGroup>
-                    </Command>
-                  </PopoverContent>
-                </Popover>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-          <FormField
-            control={form.control}
-            name="programming_lang"
-            render={({ field }) => (
-              <FormItem className="flex flex-col rounded-lg border p-4 space-y-1">
-                <FormLabel>Favorite Language</FormLabel>
-                <FormDescription>
-                  Pick your most proficient language.
-                </FormDescription>
-                <Popover>
-                  <PopoverTrigger asChild>
-                    <FormControl>
-                      <Button
-                        variant="outline"
-                        role="combobox"
-                        className={cn(
-                          "w-[200px] justify-between",
-                          !field.value &&
-                            "text-muted-foreground text-devready-green"
-                        )}
-                      >
-                        {field.value
-                          ? programmingLanguagesList.find(
-                              (lang) => lang.label === field.value
-                            )?.label
-                          : "Select language"}
-                        <CheckCheckIcon className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-                      </Button>
-                    </FormControl>
-                  </PopoverTrigger>
-                  <PopoverContent className="w-[200px] p-0">
-                    <Command className=" overflow-y-auto max-h-[420px]">
-                      <CommandInput placeholder="Search language..." />
-                      <CommandEmpty>No language found.</CommandEmpty>
-                      <CommandGroup>
-                        {programmingLanguagesList.map((lang) => (
-                          <CommandItem
-                            value={lang.label}
-                            key={lang.label}
-                            onSelect={() => {
-                              form.setValue("programming_lang", lang.label);
-                            }}
-                          >
-                            <CheckIcon
-                              className={cn(
-                                "mr-2 h-4 w-4",
-                                lang.label === field.value
-                                  ? "opacity-100"
-                                  : "opacity-0"
-                              )}
-                            />
-                            {lang.label}
-                          </CommandItem>
-                        ))}
-                      </CommandGroup>
-                    </Command>
-                  </PopoverContent>
-                </Popover>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-        </div>
-        {/* Career Level */}
-        <FormField
-          control={form.control}
-          name="career_level"
-          render={({ field: { value, onChange } }) => (
-            <FormItem className="space-y-1">
-              <div className="flex flex-col items-center md:items-start pb-2">
-                <FormLabel>Professional Stage</FormLabel>
-                <FormDescription>
-                  Indicate your current professional stage.
-                </FormDescription>
-              </div>
-
-              <FormControl className="mx-4 w-11/12">
-                <Slider
-                  min={0}
-                  max={100}
-                  step={1}
-                  defaultValue={[defaultValues?.career_level || 0]}
-                  onValueChange={(vals) => {
-                    onChange(vals[0]);
-                  }}
-                />
-              </FormControl>
-
-              <div className="flex justify-between text-xs text-muted-foreground ml-1 sm:pr-8 md:pr-10 lg:pr-16 xl:pr-12">
-                <div className="flex flex-col text-center">
-                  <p>Aspiring</p>
-                  <p className="w-full">Dev</p>
-                </div>
-                <div className="flex flex-col items-center text-center">
-                  <p>Junior</p>
-                  <p>Dev</p>
-                </div>
-                <div className="flex flex-col items-center text-center">
-                  <p>Mid-Level</p>
-                  <p>Dev</p>
-                </div>
-                <div className="flex flex-col items-end text-center">
-                  <p>Senior</p>
-                  <p className="w-full">Dev</p>
-                </div>
-              </div>
-
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        {/* Experience level */}
-        <FormField
-          control={form.control}
-          name="experience_level"
-          render={({ field: { value, onChange } }) => (
-            <FormItem className="space-y-1">
-              <div className="flex flex-col items-center md:items-start pb-2">
-                <FormLabel>Experience Level</FormLabel>
-                <FormDescription>
-                  Indicate your experience in years of work.
-                </FormDescription>
-              </div>
-
-              <FormControl className="mx-4 w-11/12">
-                <Slider
-                  min={0}
-                  max={100}
-                  step={1}
-                  defaultValue={[defaultValues?.experience_level || 0]}
-                  onValueChange={(vals) => {
-                    onChange(vals[0]);
-                  }}
-                />
-              </FormControl>
-
-              <div className="flex justify-between text-xs text-muted-foreground ml-1 sm:pr-8 md:pr-10 lg:pr-16 xl:pr-12">
-                <div className="flex flex-col text-center">
-                  <p>&gt;6</p>
-                  <p className="w-full">Months</p>
-                </div>
-                <div className="flex flex-col items-center text-center">
-                  <p>1</p>
-                  <p>Year</p>
-                </div>
-                <div className="flex flex-col items-center text-center">
-                  <p>3</p>
-                  <p>Years</p>
-                </div>
-                <div className="flex flex-col items-end text-center">
-                  <p>5+</p>
-                  <p className="w-full">Years</p>
-                </div>
-              </div>
-
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        {/* Skills 🎯 */}
-        <FormField
-          control={form.control}
-          name="skills_list"
-          render={({ field: { value, onChange } }) => (
-            <FormItem className="space-y-1">
-              <div className="flex flex-col items-center text-center md:text-start md:items-start pb-2">
-                <FormLabel>Pick your proficient skills</FormLabel>
-                <FormDescription>
-                  Pick your most proficient skill set
-                </FormDescription>
-              </div>
-              <FormControl>
-                <ToggleGroup
-                  size={"sm"}
-                  variant="skill"
-                  type="multiple"
-                  className="flex flex-row flex-wrap"
-                  aria-label="Skills list"
-                >
-                  {skillsList.map((skill) => (
-                    <ToggleGroupItem
-                      key={skill.label}
-                      value={skill.label.toLowerCase()}
-                      aria-label={skill.label}
-                      //🎯 need to update onCLick functionality - handle form update
-                      onClick={() => handleSkillList}
-                    >
-                      {skill.label}
-                    </ToggleGroupItem>
-                  ))}
-                </ToggleGroup>
-              </FormControl>
-            </FormItem>
-          )}
-        />
+      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
         {/* Bio  */}
         <FormField
           control={form.control}
           name="bio"
           render={({ field }) => (
             <FormItem className="space-y-1">
-              <div className="flex flex-col items-center text-center md:text-start md:items-start pb-2">
-                <FormLabel>Bio</FormLabel>
+              <div className="flex flex-col text-start pb-2">
+                <FormLabel className="text-start justify-start">Bio</FormLabel>
                 <FormDescription>
                   Give a brief description of yourself and your learning journey
                 </FormDescription>
               </div>
-              <FormControl className="flex items-center text-center sm:items-start sm:text-start">
+              <FormControl className="flex items-start sm:text-start">
                 <Textarea
                   placeholder="Tell us a little bit about yourself"
                   className="resize-none"
@@ -367,12 +121,275 @@ export function ProfileForm() {
             </FormItem>
           )}
         />
+
+        {/* Language & location */}
+        <div className="flex flex-col md:flex-row gap-3 md:gap-10 mr-16">
+          <FormField
+            control={form.control}
+            name="language"
+            render={({ field }) => (
+              <FormItem className="flex flex-col rounded-lg border p-4">
+                <FormLabel>Language</FormLabel>
+                <Popover>
+                  <PopoverTrigger asChild>
+                    <FormControl>
+                      <Button
+                        variant="outline"
+                        role="combobox"
+                        className={cn(
+                          "w-[200px] justify-between",
+                          !field.value &&
+                            "text-muted-foreground text-devready-green"
+                        )}
+                      >
+                        {field.value
+                          ? languages.find(
+                              (language) => language.value === field.value
+                            )?.label
+                          : "Select language"}
+                        <CheckCheckIcon className="ml-2 h-4 w-4 shrink-0 opacity-50" />
+                      </Button>
+                    </FormControl>
+                  </PopoverTrigger>
+                  <PopoverContent className="w-[200px] p-0">
+                    <Command>
+                      <CommandInput placeholder="Search language..." />
+                      <CommandEmpty>No language found.</CommandEmpty>
+                      <CommandGroup>
+                        {languages.map((language) => (
+                          <CommandItem
+                            value={language.label}
+                            key={language.value}
+                            onSelect={() => {
+                              form.setValue("language", language.value);
+                            }}
+                          >
+                            <CheckIcon
+                              className={cn(
+                                "mr-2 h-4 w-4",
+                                language.value === field.value
+                                  ? "opacity-100"
+                                  : "opacity-0"
+                              )}
+                            />
+                            {language.label}
+                          </CommandItem>
+                        ))}
+                      </CommandGroup>
+                    </Command>
+                  </PopoverContent>
+                </Popover>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="country"
+            render={({ field }) => (
+              <FormItem className="flex flex-col rounded-lg border p-4">
+                <FormLabel>Location</FormLabel>
+                <Popover>
+                  <PopoverTrigger asChild>
+                    <FormControl>
+                      <Button
+                        variant="outline"
+                        role="combobox"
+                        className={cn(
+                          "w-[200px] justify-between",
+                          !field.value &&
+                            "text-muted-foreground text-devready-green"
+                        )}
+                      >
+                        {field.value
+                          ? countriesList.find(
+                              (country) => country.value === field.value
+                            )?.label
+                          : "Select country"}
+                        <CheckCheckIcon className="ml-2 h-4 w-4 shrink-0 opacity-50" />
+                      </Button>
+                    </FormControl>
+                  </PopoverTrigger>
+                  <PopoverContent className="w-[200px] p-0">
+                    <Command className=" overflow-y-auto max-h-[420px]">
+                      <CommandInput placeholder="Search country..." />
+                      <CommandEmpty>No country found.</CommandEmpty>
+                      <CommandGroup className="overflow-y-auto max-h-[300px]">
+                        {countriesList.map((country) => (
+                          <CommandItem
+                            value={country.label}
+                            key={country.value}
+                            onSelect={() => {
+                              form.setValue("country", country.value);
+                            }}
+                            // className=" overflow-y-auto max-h-[10px]"
+                          >
+                            <CheckIcon
+                              className={cn(
+                                "mr-2 h-4 w-4",
+                                country.value === field.value
+                                  ? "opacity-100"
+                                  : "opacity-0"
+                              )}
+                            />
+                            {country.label}
+                          </CommandItem>
+                        ))}
+                      </CommandGroup>
+                      {/* <h1 className="ml-10 text-primary/40">
+                        <MoreHorizontal />
+                      </h1> */}
+                    </Command>
+                  </PopoverContent>
+                </Popover>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
+
+        {/* Connections */}
+        <div className="flex flex-col gap-2">
+          <FormField
+            control={form.control}
+            name="urls.github"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Connections</FormLabel>
+                <FormDescription>
+                  Connect with your community and make your profile stand out.
+                </FormDescription>
+                <div className="flex flex-row justify-between items-center gap-3">
+                  <Github className="text-devready-green" size={20} />
+                  <FormControl className="w-full">
+                    <Input placeholder="GitHub profile" {...field} />
+                  </FormControl>
+                </div>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="urls.linkedin"
+            render={({ field }) => (
+              <FormItem>
+                <div className="flex flex-row justify-between items-center gap-3">
+                  <Linkedin className="text-devready-green" size={20} />
+                  <FormControl className="w-full">
+                    <Input placeholder="LinkedIn profile" {...field} />
+                  </FormControl>
+                </div>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="urls.website"
+            render={({ field }) => (
+              <FormItem>
+                <div className="flex flex-row justify-between items-center gap-3">
+                  <LinkIcon className="text-devready-green" size={20} />
+                  <FormControl className="w-full">
+                    <Input placeholder="Portfolio or project" {...field} />
+                  </FormControl>
+                </div>
+
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
+        {/* Projects */}
+        <div className="flex flex-col gap-2">
+          <FormField
+            control={form.control}
+            name="projects.capstone"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Projects</FormLabel>
+                <FormDescription>
+                  Showcase your most proudest projects.
+                </FormDescription>
+                <div className="flex flex-row justify-between items-center gap-3">
+                  <Trophy className="text-devready-green" size={20} />
+                  <FormControl className="w-full">
+                    <Input placeholder="Capstone project" {...field} />
+                  </FormControl>
+                </div>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="projects.additional"
+            render={({ field }) => (
+              <FormItem>
+                <div className="flex flex-row justify-between items-center gap-3">
+                  <Medal className="text-devready-green" size={20} />
+                  <FormControl className="w-full">
+                    <Input placeholder="Additional project" {...field} />
+                  </FormControl>
+                </div>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
+        {/* Switches */}
+        <div className="flex flex-col gap-2">
+          <FormField
+            control={form.control}
+            name="ztm_student"
+            render={({ field }) => (
+              <FormItem className="flex flex-row items-center justify-between rounded-lg border p-3">
+                <div className="">
+                  <FormLabel className="text-base">ZTM Alumni</FormLabel>
+                  <FormDescription>
+                    Are you currently a ZTM student or alumni?
+                  </FormDescription>
+                </div>
+                <FormControl>
+                  <Switch
+                    checked={field.value}
+                    onCheckedChange={field.onChange}
+                    // disabled
+                    // aria-readonly
+                  />
+                </FormControl>
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="star_mentor"
+            render={({ field }) => (
+              <FormItem className="flex flex-row items-center justify-between rounded-lg border p-3">
+                <div className="">
+                  <FormLabel className="text-base">Star Mentor</FormLabel>
+                  <FormDescription>Are you a star mentor?</FormDescription>
+                </div>
+                <FormControl>
+                  <Switch
+                    checked={field.value}
+                    onCheckedChange={field.onChange}
+                    disabled
+                    aria-readonly
+                  />
+                </FormControl>
+              </FormItem>
+            )}
+          />
+        </div>
+
         <Button
           type="submit"
           variant={"devfill"}
           className="rounded-lg text-sm md:text-sm p-2"
         >
-          Update profile
+          Update account
         </Button>
       </form>
     </Form>
